requirements.txt
django
elasticsearch
django-elasticsearch-dsl
djangorestframework
requests